avm2: Move vector class initialization code from globals.rs into vector.rs
